+++ src/ContaoManager/Plugin.php
@@ -17,7 +17,7 @@ class Plugin implements BundlePluginInterface
 	{
 		return [
 			BundleConfig::create(ContaoSpielerregisterBundle::class)
-				->setLoadAfter([ContaoCoreBundle::class]),
+				->setLoadAfter([ContaoCoreBundle::class, ContaoNewsletterBundle::class]),
 		];
 	}
 }