+++ src/Resources/contao/languages/de/tl_newsletter_recipients.php
@@ -1,0 +1,13 @@
+<?php 
+
+/**
+ * Contao Open Source CMS
+ * Copyright (C) 2005-2014 Leo Feyer
+ *
+ */
+
+
+/**
+ * Back end modules
+ */
+$GLOBALS['TL_LANG']['tl_newsletter_recipients']['spielerregister_mailTime'] = array('Versandzeit Newsletter', 'Versandzeit des letzten Spielerregister-Newsletters');